requirements.txt
requests
PyYAML
PyQt5
numpy
